db/requirements_db.txt
numpy==1.24.4 
pandas==1.5.3
kaggle==1.5.12
psycopg2-binary==2.9.5